changed current count to u64

+++ solana-project/programs/solana-project/src/events.rs
@@ -27,25 +27,25 @@ pub struct RegisteredChain {
 pub struct StoredMsg {
     pub msg_type: u64,
     pub sender: [u8; 32],
-    pub count: u8,
+    pub count: u64,
 }
 
 #[event]
 pub struct Deposited {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct StreamUpdated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct PausedResumed {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
@@ -54,7 +54,7 @@ pub struct DirectTransferredNative {
     pub sender_chain: Vec<u8>,
     pub target_chain: u16,
     pub receiver: Vec<u8>,
-    pub current_count: u8
+    pub current_count: u64
 }
 
 #[event]
@@ -63,37 +63,37 @@ pub struct DirectTransferredWrapped {
     pub sender_chain: Vec<u8>,
     pub target_chain: u16,
     pub receiver: Vec<u8>,
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct StreamCreated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct CancelCreated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct SenderWithdrawCreated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct InstantTransferCreated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
 pub struct ReceiverWithdrawCreated {
     pub sender: [u8; 32],
-    pub current_count: u8,
+    pub current_count: u64,
 }
 
 #[event]
